myhotel/catalog/serializers.py
from rest_framework import serializers
from .models import Room


class RoomSerializer(serializers.ModelSerializer):
    class Meta:
        model = Room
        fields = ['room_class', 'bed_type', 'price_per_night', 'description']